fix: type of overall rating for book entity

+++ src/main/java/com/bootcamp/bookrentalsystem/service/ReviewService.java
@@ -14,10 +14,8 @@ import java.util.UUID;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Sort;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
@@ -95,9 +93,9 @@ public class ReviewService {
 
         reviewRepository.save(review);
 
-        Map<Double, Integer> result = calculateOverallRating(review.getBookId());
+        Map<Integer, Integer> result = calculateOverallRating(review.getBookId());
 
-        double updateBookOverAllRating = result.keySet().iterator().next();
+        int updateBookOverAllRating = result.keySet().iterator().next();
         int numberOfReviews = result.get(updateBookOverAllRating);
         book.setOverAllRating(updateBookOverAllRating);
         book.setReviewsCount(numberOfReviews);
@@ -133,8 +131,8 @@ public class ReviewService {
             }
             existingReview.setRating(rating);
             // Recalculate the Book overall rating
-            Map<Double, Integer> result = calculateOverallRating(existingReview.getBookId());
-            double updateBookOverAllRating = result.keySet().iterator().next();
+            Map<Integer, Integer> result = calculateOverallRating(existingReview.getBookId());
+            int updateBookOverAllRating = result.keySet().iterator().next();
             int numberOfReviews = result.get(updateBookOverAllRating);
 
             // Save the update book
@@ -167,8 +165,8 @@ public class ReviewService {
 
         reviewRepository.delete(existingReview);
         // Recalculate the Book overall rating
-        Map<Double, Integer> result = calculateOverallRating(existingReview.getBookId());
-        double updateBookOverAllRating = result.keySet().iterator().next();
+        Map<Integer, Integer> result = calculateOverallRating(existingReview.getBookId());
+        int updateBookOverAllRating = result.keySet().iterator().next();
         int numberOfReviews = result.get(updateBookOverAllRating);
 
         // Save the update book
@@ -188,13 +186,13 @@ public class ReviewService {
         return reviewRepository.findByBookId(bookId, sort);
     }
 
-    public Map<Double, Integer> calculateOverallRating(UUID bookId) {
+    public Map<Integer, Integer> calculateOverallRating(UUID bookId) {
         bookRepository.findById(bookId)
                 .orElseThrow(() -> new ResourceNotFoundException("Book not found with ID: " + bookId));
         List<Review> reviews = getAllReviewsByBook(bookId);
         if (reviews == null || reviews.isEmpty()) {
-            Map<Double, Integer> result = new HashMap<>();
-            result.put(0.0, 0);
+            Map<Integer, Integer> result = new HashMap<>();
+            result.put(0, 0);
             return result;
         }
 
@@ -204,9 +202,9 @@ public class ReviewService {
         for (Review review : reviews) {
             totalRatings += review.getRating();
         }
-        double averageRating = (double) totalRatings / numberOfReviews;
+        int averageRating = (int) totalRatings / numberOfReviews;
 
-        Map<Double, Integer> result = new HashMap<>();
+        Map<Integer, Integer> result = new HashMap<>();
         result.put(averageRating, numberOfReviews);
         return result;
